Form validation and updates

+++ index.php
@@ -7,7 +7,6 @@ $i = 1;
 
 include 'partials/header.php';
 
-if ($result->num_rows > 0) {
     echo "<div class='top'>";
     echo "<h2>Contact List</h2>";
     echo "<a href='create.php' class='create-link'>Create a New Contact</a>";
@@ -25,19 +24,24 @@ if ($result->num_rows > 0) {
                     <th> </th>
                 </tr>";
     // Output data of each row
-    while ($row = $result->fetch_assoc()) {
-        $id = $row["id"];  // Assuming 'id' is the correct column name for the primary key in your database
-        echo "<tr>
-                    <td>" . $i . "</td>
-                    <td><a href='view_details.php?id=" . $id . "' class='name-link'>". $row["contact_Name"] . "</a> </td>
-                    <td>" . $row["phone"] . "</td>
-                    <td><a href='edit.php?id=" . $id . "' class='btn-link'>Edit</a></td>
-                    <td><a href='delete.php?id=" . $id . "' class='btn-link' onclick='return confirm(\"Are you sure you want to delete this contact?\");'>Delete</a></td>
-                  </tr>";
-        $i++;
+    if ($result->num_rows > 0) {
+        while ($row = $result->fetch_assoc()) {
+            $id = $row["id"];  // Assuming 'id' is the correct column name for the primary key in your database
+            echo "<tr>
+                        <td>" . $i . "</td>
+                        <td><a href='view_details.php?id=" . $id . "' class='name-link'>". $row["contact_Name"] . "</a> </td>
+                        <td>" . $row["phone"] . "</td>
+                        <td><a href='edit.php?id=" . $id . "' class='btn-link'>
+                        <span class='material-symbols-outlined' title='Edit Contact'>edit_square</span>
+                        </a></td>
+                        <td><a href='delete.php?id=" . $id . "' class='btn-link' id='delete' onclick='return confirm(\"Are you sure you want to delete this contact?\");'> 
+                         <span class='material-symbols-outlined' title='Delete Contact'>delete</span>
+                        </a></td>
+                    </tr>";
+            $i++;
+        }
+        echo "</table>";
+    } else {
+        echo "No results!";
     }
-    echo "</table>";
-} else {
-    echo "No results!";
-}
 include 'partials/footer.php';
